Correct commit messages for given diff:
"Add condition for year in FormattedDate component"

The FormattedDate component has been updated to only display the date when the year is not equal to 1900. This change was made to prevent displaying incorrect or irrelevant dates. The date in 'FormattedDate.astro' will now only be displayed when the 'getFullYear()' method returns a value different from 1900. This ensures that only meaningful and accurate dates are presented to the user.