Remove third-party CSS attribution in global.css
The CSS attribution that made reference to the source code from Bear Blog's default CSS is removed from global.css file. This is done to declutter the file and avoid any confusion as our project does not rely on the external CSS. The removal of attribution allows for cleaner and more maintainable code.