Update site URL in astro.config
Changed the site URL in the astro.config.mjs from 'https://example.com' to 'https://dotnetevangelist.net'. The change was made to accurately represent the correct website within the configuration file. The correct URL ensures that any generated sitemaps or other data based on this config file will point to the right location.