Add 'pagefind-modular-ui.js' and 'pagefind.js' files
The 'pagefind-modular-ui.js' and 'pagefind.js' have been added to the public directory. The former is a modular user interface JavaScript file while the latter is a general JavaScript file for searching functions. Both files are integral to the functionality of the application, providing UI components, event handlers and search capabilities.